[Keymap] RGB Timeout added to userspace (#13339)
